fix: codacy errors
